requirements.txt
requests
PyPDF2
Pillow
beautifulsoup4
PyQt5
pyqt-toast-notification
img2pdf
pygame
selenium
cloudscraper
PySide6>=6.5.1
Pillow>=5.2.0
psutil>=5.9.5
requests>=2.31.0
python-slugify>=1.2.1
raven>=6.0.0
mozjpeg-lossless-optimization>=1.1.2
natsort[fast]>=8.4.0
distro>=1.8.0